Modify App.js
Moved the logic of the application to App.js : does the main computation
and transfer the results to child components.
Add a new Calendar component in children.
Data is now fetched when the calendar finish getting current week. This
will allow in the future to directly select week in calendar.
App.js also handles the clicks on a day in calendar and transfer the
selected day to the list of offers.